messages: make `sig` take a mapping and batch it
Not batching it was unnecessary (to say the least), and to force the
watchtower to use public key recovery was imposing an unnecessary
complexity on it.

Using a mapping is also more natural given our usage of the Miniscript
satisfier.

We are now quite explicit: we signal the vault id and the txid these
signatures are for.

We also remove the useless txid obfuscation as there are now middle-man
between a wallet and its watchtower.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>