messages: add the chosen fixed feerate to the unvault transaction
To sum up, we cannot broadcast the unvault if the majority of the network mempools
minfee goes above 21sat/vbyte, but there is no fix apart from deploying
package relay on the Bitcoin network (https://github.com/bitcoin/bitcoin/issues/14895).

This is not security critical, but may become so as users may be trying
any trick to bypass the blocking of operations.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>