messages: encrypt the emergency transaction signature
This extends the `sig` message with a new optional `encrypted_signature`
entry, which if set voids the `signature` entry. This is in order to
avoid introducing a new message for the emergency transaction.

We can reasonnably assume that the "remove the other entry" trick will
not introduce a convoluted logic in the implementation as callers know
the type of the transaction before requesting it.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>