transactions: add an OP_RETURN output to prevent input replay
Due to exchanging ALL | ANYONECANPAY signatures, Unvault and Deposit
inputs could be replayed in other Cancel or Emergency transactions,
effectively merging those transactions.

Even though unlikely and in most cases requiring internal corruption
this would be a massive DOS, so prevent them by comitting to specific
vaults.

There exists more complicated way of comitting to each vault but we go
with a simple-to-reason-about solution for now.

Reported-by: Stepan Snigirev <snigirev.stepan@gmail.com>
Signed-off-by: Antoine Poinsot <darosior@protonmail.com>