messages: make the cosigning server return previous sigs
In the first place, it was a mistake to not store the one time
signatures on the server side.
Now, it's actually very useful on the happy case where a client just
wants to retry the same Spend transaction because for some reason the
response was dropped on the road to them.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>