transactions: increase Emergency feerate to 300sat/vb
Fixes #86, although it's not an "insanely high feerate" as mentioned in
the issue.

The rationale for this was that WTs won't have a fee reserve dedicated
to the Emergency fee bumping. We can use the reserve kept by watchtowers
for bumping the Cancel, but it's only kept for active vaults.
300sat/vb is meant to be a sweet pot between:
- too much reliance on WTs' reserves for the Cancel tx that they may not
  have (eg if most vaults are inactive)
- incurring a too high barrier to entry (a larger feerate increases the
  minimum deposit value) and an Emergency cost (what % of funds are you
  ready to burn during an Emergency procedure? Supposedly 100% for game
  theory but hey)