introduction: replace out-dated revault.pdf with introduction.md
Removed:

The 'Keys' section was removed in favour of a more detailed spec coming with #107
The 'Transactions' section was removed as it duplicates transactions.md
'Acknowledgements' section removed. Should it be somewhere else in the repo?

New content:

Updated transaction diagram using descriptors abstraction.
Discussed options for revault deployments and the infrastructure requriements.
Renamed "Threat model" section to "Risk" and updated.
The 'fee bumping' section was mostly removed as we are approaching it differently.
'Further Discussion' removed as this information is mostly out-dated or duplicated.