Presign all transactions with SIGHASH_ALL, remove ANYONECANPAY mentions
There are still too many pinning attacks possible, such that it's not
reasonable to deploy with this. In addition, it's certainly not worth
the complexity and overhead of managing fee-bumping reserves.

This is in theory a reduction in security as the maximum reserve is now
a deployment parameter (at what feerate to sign the Cancel txs) where it
was previously flexible and left to the discretion of each watchtower.