transactions: presign the Emergency transaction at 1000sat/vb
Since we got rid of fee-bumping, we need a feerate such that it is
*very* unlikely the next block feerate ever goes above (which'd make the
deterrent somewhat moot).

On the other hand, choosing something much higher really conditions the
deposit size. Especially in case of deposit split.
For 8 stakeholders and 3 managers, this makes the deposit be at least
500€ in order for the Emergency to not burn >25% of the funds at this
feerate. (Emergency vbytes: 211, BTCEUR: 40k.)